% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.R
\name{aic}
\alias{aic}
\title{Calculates the AICs}
\usage{
aic(qp)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{aic} calculates the AICs.
}
\examples{
library(MPDiR) # contains the Vernier data
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq), B = 20)
aic(fit)
}

