% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akaike.R
\name{akaike}
\alias{akaike}
\title{Calculates the Akaike Information Criterion
\code{akaike} calculates the loglikelihoods.}
\usage{
akaike(logliks)
}
\description{
Calculates the Akaike Information Criterion
\code{akaike} calculates the loglikelihoods.
}
