% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avbootstrap.R
\name{avbootstrap}
\alias{avbootstrap}
\title{Creates bootstrap samples}
\usage{
avbootstrap(qp, bootstrap = "parametric", B = 100)
}
\arguments{
\item{qp}{output from quickpsy}

\item{bootstrap}{\code{'parametric'} performs parametric bootstrap;
\code{'nonparametric'} performs non-parametric bootstrap;
\code{'none'} does not perform bootstrap (default is \code{'parametric'}).}

\item{B}{number of bootstrap samples (default is 100 ONLY).}
}
\description{
\code{avbootstrap} creates bootstrap samples
}
\examples{
library(MPDiR) # contains the Vernier data
data(Vernier) # ?Venier for the reference
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq),
                bootstrap = 'none')
boot_samples <- parbootstrap(fit)
head(boot_samples)
}

