% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averages.R
\name{averages}
\alias{averages}
\title{Perform the averages
\code{averages} perform the averages}
\usage{
averages(d, x, x_str, k, n, log, grouping)
}
\description{
Perform the averages
\code{averages} perform the averages
}
\keyword{internal}
