% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_parini.R
\name{calculate_parini}
\alias{calculate_parini}
\title{\code{calculate_parini} calculate_parini}
\usage{
calculate_parini(averages, funname, x, guess, lapses, grouping)
}
\description{
\code{calculate_parini} calculate_parini
}
\keyword{internal}
