% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_full_nll.R
\name{create_full_nll}
\alias{create_full_nll}
\title{Creates the full negative log-likelihood function
\code{create_full_nll} Creates the full negative log-likelihood function}
\usage{
create_full_nll(d, x, k, n, psyfunguesslapses)
}
\description{
Creates the full negative log-likelihood function
\code{create_full_nll} Creates the full negative log-likelihood function
}
\keyword{internal}

