% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_nll_cum_norm}
\alias{create_nll_cum_norm}
\title{It creates from the data the negative log-likelihood as a function of the parameters.}
\usage{
create_nll_cum_norm(d, x, k, n, guess, lapses)
}
\arguments{
\item{d}{The data frame containing the data.}

\item{x}{Name of the independent variable.}

\item{k}{Number of 'yes' or correct responses.}

\item{n}{Number of trials.}

\item{psy_fun}{Function to create the negative log-likelihood.}
}
\value{
It returns a function. The negative log-likelihood as a function of the parameters (p).
}
\description{
It creates from the data the negative log-likelihood as a function of the parameters.
}

