% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nll_saturated.R
\name{create_nll_saturated}
\alias{create_nll_saturated}
\title{Creates the saturated negative log-likelihood function
\code{create_nll} Creates the saturatednegative log-likelihood function}
\usage{
create_nll_saturated(averages)
}
\description{
Creates the saturated negative log-likelihood function
\code{create_nll} Creates the saturatednegative log-likelihood function
}
\keyword{internal}
