% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nllsaturated.R
\name{create_nllsaturated}
\alias{create_nllsaturated}
\title{Creates the full saturated negative log-likelihood function
\code{create_nll} Creates the full saturated negative log-likelihood function}
\usage{
create_nllsaturated(d, x, k, n, psyfunguesslapses)
}
\description{
Creates the full saturated negative log-likelihood function
\code{create_nll} Creates the full saturated negative log-likelihood function
}
\keyword{internal}

