% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/curvesbootstrap.R
\name{curvesbootstrap}
\alias{curvesbootstrap}
\title{Creates the bootstrap curves
\code{curvesbootstrap} creates the bootstrap curves}
\usage{
curvesbootstrap(qp, xmin = NULL, xmax = NULL, log = F)
}
\description{
Creates the bootstrap curves
\code{curvesbootstrap} creates the bootstrap curves
}
\keyword{internal}

