% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devi.R
\name{devi}
\alias{devi}
\title{Calculates the deviance}
\usage{
devi(logliks, loglikssaturated, grouping_without_fun)
}
\arguments{
\item{logliks}{The \code{logliks} data frame from quickpsy.}

\item{loglikssaturated}{The \code{loglikssaturated} data frame from quickpsy.}

\item{grouping_without_fun}{The \code{grouping_without_fun} data frame from quickpsy.}
}
\description{
\code{devi} calculates the deviance
}
