% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance}
\alias{deviance}
\title{Calculates the deviances}
\usage{
deviance(qp)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{deviance} calculates the deviances.
}
\examples{
library(MPDiR) # contains the Vernier data
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq), B = 20)
deviance(fit)
}

