% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devianceboot.R
\name{devianceboot}
\alias{devianceboot}
\title{Calculates the bootsrap deviances}
\usage{
devianceboot(qp)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{deviance} calculates the bootstrap deviances.
}
\examples{
library(MPDiR) # contains the Vernier data
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq), B = 20)
devianceboot(fit)
}

