% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviancep.R
\name{deviancep}
\alias{deviancep}
\title{Calculates the probability of obtaining the deviance value or larger if the parametric model holds
\code{deviancep} calculates the probability of obtaining the deviance value or larger if the parametric model holds.}
\usage{
deviancep(qp)
}
\description{
Calculates the probability of obtaining the deviance value or larger if the parametric model holds
\code{deviancep} calculates the probability of obtaining the deviance value or larger if the parametric model holds.
}
\keyword{internal}

