% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpsy.R
\name{fitpsy}
\alias{fitpsy}
\title{Fits the curve
\code{fitpsy} fits de curve}
\usage{
fitpsy(d, x, k, n, grouping, xmin, xmax, log, fun, funname, parini, pariniset,
  guess, lapses, optimization)
}
\description{
Fits the curve
\code{fitpsy} fits de curve
}
\keyword{internal}
