% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitpsyBackUp2.R
\name{fitpsybackup2}
\alias{fitpsybackup2}
\title{fitpsy}
\usage{
fitpsybackup2(d = data, x = x, k = response, n = NULL, random, within,
  between, xmin = NULL, xmax = NULL, log = F, psyfun = cum_normal_fun,
  pini = NULL, guess = 0, lapses = 0, DE = F, pini2 = NULL)
}
\description{
fitpsy
}

