% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpsys.R
\name{fitpsys}
\alias{fitpsys}
\title{Fits the curves
\code{fitpsys} fits de curves}
\usage{
fitpsys(d, x, k, n, random, within, between, grouping, xmin, xmax, log, funname,
  parini, optimization)
}
\description{
Fits the curves
\code{fitpsys} fits de curves
}
\keyword{internal}

