% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_functions}
\alias{get_functions}
\title{Predefined functions}
\usage{
get_functions()
}
\description{
\code{getfunctions} lists the predefined functions in \code{quickpsy}.
}
\seealso{
\code{\link{cum_normal_fun}},
\code{\link{logistic_fun}},
\code{\link{weibull_fun}}
}

