% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logistic_fun.R
\name{inv_logistic_fun}
\alias{inv_logistic_fun}
\title{Inverse logistic  function}
\usage{
inv_logistic_fun(q, p)
}
\arguments{
\item{q}{Vector of probabilities.}

\item{p}{Vector of parameters \eqn{p = c(\alpha, \beta)}.}
}
\value{
\code{x} at each probability.
}
\description{
Inverse logistic function
}
\examples{
yseq <- seq(0, 1, .01)
xseq <- inv_logistic_fun(yseq, c(2, 4))
curve <- data.frame(x = xseq, y = yseq)
ggplot(curve, aes(x = x, y = y)) + geom_line()
}
\seealso{
\code{\link{logistic_fun}}
}

