% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_weibull_fun.R
\name{inv_weibull_fun}
\alias{inv_weibull_fun}
\title{Inverse Weibull  function}
\usage{
inv_weibull_fun(q, p)
}
\arguments{
\item{q}{Vector of probabilities.}

\item{p}{Vector of parameters p = c(alpha, beta).}
}
\value{
\code{x} at each probability.
}
\description{
Inverse Weibull function
}
\examples{
yseq <- seq(0, 1, .01)
xseq <- inv_weibull_fun(yseq, c(2, 4))
curve <- data.frame(x = xseq, y = yseq)
ggplot(curve, aes(x = x, y = y)) + geom_line()
}
\seealso{
\code{\link{weibull_fun}}
}

