% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{limits}
\alias{limits}
\title{Creates the limits
\code{limits} creates the limits}
\usage{
limits(averages, x_str, xmin, xmax)
}
\description{
Creates the limits
\code{limits} creates the limits
}
\keyword{internal}
