% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliks.R
\name{logliks}
\alias{logliks}
\title{Calculates the loglikelihoods
\code{logliks} calculates the loglikelihoods.}
\usage{
logliks(nll_fun, param, grouping_without_fun)
}
\arguments{
\item{nll_fun}{The \code{nll_fun} data frame from quickpsy.}

\item{param}{The \code{param} data frame from quickpsy.}

\item{grouping_without_fun}{The \code{grouping_without_fun} data frame from quickpsy.}
}
\description{
Calculates the loglikelihoods
\code{logliks} calculates the loglikelihoods.
}
