% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliks_saturated.R
\name{logliks_saturated}
\alias{logliks_saturated}
\title{Calculates the saturated loglikelihoods
\code{logliks} calculates the saturatedloglikelihoods.}
\usage{
logliks_saturated(nll_fun_saturated, averages)
}
\description{
Calculates the saturated loglikelihoods
\code{logliks} calculates the saturatedloglikelihoods.
}
