% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliksboot.R
\name{logliksboot}
\alias{logliksboot}
\title{Calculates the bootstrap loglikelihoods}
\usage{
logliksboot(qp)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{logliksboot} calculates the bootstraploglikelihoods.
}
\examples{
library(MPDiR) # contains the Vernier data
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq), B = 20)
logliksboot(fit)
}

