% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliksbootsaturated.R
\name{logliksbootsaturated}
\alias{logliksbootsaturated}
\title{Calculates the bootstrap loglikelihoods for the saturated model}
\usage{
logliksbootsaturated(qp)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{logliks} calculates the bootstrap loglikelihoods for the saturated model.
}
\examples{
library(MPDiR) # contains the Vernier data
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq), B = 20)
logliksbootsaturated(fit)
}

