% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikssaturated.R
\name{loglikssaturated}
\alias{loglikssaturated}
\title{Calculates the saturated loglikelihoods
\code{logliks} calculates the saturated loglikelihoods.}
\usage{
loglikssaturated(nll_fun, param)
}
\description{
Calculates the saturated loglikelihoods
\code{logliks} calculates the saturated loglikelihoods.
}
