% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_aic.R
\name{model_selection_aic}
\alias{model_selection_aic}
\title{Select models using the Akaike Information Criterion}
\usage{
model_selection_aic(aic1, aic2)
}
\arguments{
\item{aic1}{The \code{aic1} data frame from quickpsy for the first model.}

\item{aic2}{The \code{aic2} data frame from quickpsy for the second model.}
}
\description{
The \code{p} indicates the relative probability of the models.
}
