% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_lrt.R
\name{model_selection_lrt}
\alias{model_selection_lrt}
\title{Select models using the likelihood ratio test}
\usage{
model_selection_lrt(loglik1, loglik2, alpha = 0.05)
}
\arguments{
\item{loglik1}{The \code{loglik} data frame from quickpsy for the first model.}

\item{loglik2}{The \code{loglik} data frame from quickpsy for the second model.}

\item{alpha}{The significance level.}
}
\description{
Select models using the likelihood ratio test
}
