% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_fun_saturated.R
\name{nll_fun_saturated}
\alias{nll_fun_saturated}
\title{\code{nll_fun} Creates the sturated negative log-likelihood function}
\usage{
nll_fun_saturated(averages, psych_fun, grouping_without_fun)
}
\description{
\code{nll_fun} Creates the sturated negative log-likelihood function
}
\keyword{internal}
