% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_aic.R
\name{one_aic}
\alias{one_aic}
\title{Calculates the AIC for one condition
\code{one_aic} creates the deviance for one condition
one_deviance}
\usage{
one_aic(d, groups, par)
}
\description{
Calculates the AIC for one condition
\code{one_aic} creates the deviance for one condition
one_deviance
}
\keyword{internal}

