% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/one_bootstrap.R
\name{one_bootstrap}
\alias{one_bootstrap}
\title{one_bootstrap}
\usage{
one_bootstrap(d, x, k, n, psyfunguesslapses, funname, guess, lapses, pini,
  pini2, bootstrap, B, groups, ypred)
}
\description{
one_bootstrap
}

