% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/one_curve.R
\name{one_curve}
\alias{one_curve}
\title{Creates the curve for one condition
\code{one_curve} Creates the curve for one condition}
\usage{
one_curve(d, xmin, xmax, log, groups, limits, psyfunguesslapses)
}
\description{
Creates the curve for one condition
\code{one_curve} Creates the curve for one condition
}
\keyword{internal}

