% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_deviance.R
\name{one_deviance}
\alias{one_deviance}
\title{Creates the deviance for one condition
\code{one_deviance} creates the deviance for one condition
one_deviance}
\usage{
one_deviance(d, groups, loglikssaturated)
}
\description{
Creates the deviance for one condition
\code{one_deviance} creates the deviance for one condition
one_deviance
}
\keyword{internal}

