% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_deviancep.R
\name{one_deviancep}
\alias{one_deviancep}
\title{Calculates the probability of obtaining the deviance value or larger if the
parametric model holds for one condition
\code{one_deviancep} calculates the probability of obtaining the deviance value
or larger if the parametric model holds for one condition}
\usage{
one_deviancep(d, groups, deviance)
}
\description{
Calculates the probability of obtaining the deviance value or larger if the
parametric model holds for one condition
\code{one_deviancep} calculates the probability of obtaining the deviance value
or larger if the parametric model holds for one condition
}
\keyword{internal}

