% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_limit.R
\name{one_limit}
\alias{one_limit}
\title{Creates the limits for one condition
\code{one_limit} creates the limits for one condition}
\usage{
one_limit(d, x, xmin, xmax, log)
}
\description{
Creates the limits for one condition
\code{one_limit} creates the limits for one condition
}
\keyword{internal}

