% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_logliksaturated.R
\name{one_logliksaturated}
\alias{one_logliksaturated}
\title{Creates the saturated loglik for one condition
\code{one_logliksaturated} creates the saturated loglik for one condition
one_loglik}
\usage{
one_logliksaturated(d, x, k, n, psyfunguesslapses, groups, par)
}
\description{
Creates the saturated loglik for one condition
\code{one_logliksaturated} creates the saturated loglik for one condition
one_loglik
}
\keyword{internal}

