% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_parameters.R
\name{one_parameters}
\alias{one_parameters}
\title{Obtains the parameters for one condition
\code{one_parameters} obtains the parameters for one condition}
\usage{
one_parameters(d, x, k, n, psyfunguesslapses, funname, parini, pariniset, guess,
  lapses, optimization, groups)
}
\description{
Obtains the parameters for one condition
\code{one_parameters} obtains the parameters for one condition
}
\keyword{internal}

