% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/one_parci.R
\name{one_parci}
\alias{one_parci}
\title{Calculates the confidence intervals for one condition
\code{one_parci} calculates the confidence intervals for one condition}
\usage{
one_parci(d, ci)
}
\description{
Calculates the confidence intervals for one condition
\code{one_parci} calculates the confidence intervals for one condition
}
\keyword{internal}

