% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/one_parcomparisons.R
\name{one_parcomparisons}
\alias{one_parcomparisons}
\title{Pair comparisons of the parameters using bootstrap (for each parameter)
\code{one_parcomparisons} Calculates the bootstrap confidence intervals for the
difference in the parameters for two groups for all possible pairs
of groups (for each parameter)}
\usage{
one_parcomparisons(d, para, groups, ci)
}
\description{
Pair comparisons of the parameters using bootstrap (for each parameter)
\code{one_parcomparisons} Calculates the bootstrap confidence intervals for the
difference in the parameters for two groups for all possible pairs
of groups (for each parameter)
}
\keyword{internal}

