% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_parini.R
\name{one_parini}
\alias{one_parini}
\title{Calculates some initial parameters
\code{parini} calculates some initial parameters}
\usage{
one_parini(d, x, k, n, guess, lapses, psyfun, groups)
}
\description{
Calculates some initial parameters
\code{parini} calculates some initial parameters
}
\keyword{internal}
