% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/one_sse.R
\name{one_sse}
\alias{one_sse}
\title{Calculates the sum of squared errors of prediction
\code{one_sse} calculates  the sum of squared errors of prediction for one
condition}
\usage{
one_sse(d, groups, averages)
}
\description{
Calculates the sum of squared errors of prediction
\code{one_sse} calculates  the sum of squared errors of prediction for one
condition
}
\keyword{internal}

