% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_thresholdcomparisons.R
\name{one_thresholdcomparisons}
\alias{one_thresholdcomparisons}
\title{Pair comparisons of the thresholds using bootstrap
\code{one_thresholdcomparisons} Calculates the bootstrap confidence intervals for the
difference in the thresholds for two groups for all possible pairs
of groups}
\usage{
one_thresholdcomparisons(d, thresholds, groups, ci)
}
\description{
Pair comparisons of the thresholds using bootstrap
\code{one_thresholdcomparisons} Calculates the bootstrap confidence intervals for the
difference in the thresholds for two groups for all possible pairs
of groups
}
\keyword{internal}

