% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/one_ypred.R
\name{one_ypred}
\alias{one_ypred}
\title{Calculate the predictions of the response variable for one condition
\code{one_ypred} calculate the predictions of the response variable for one
condition}
\usage{
one_ypred(d, log, groups, averages, x, psyfunguesslapses)
}
\description{
Calculate the predictions of the response variable for one condition
\code{one_ypred} calculate the predictions of the response variable for one
condition
}
\keyword{internal}

