% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parbootstrap.R
\name{parbootstrap}
\alias{parbootstrap}
\title{Creates bootstrap samples of the parameters}
\usage{
parbootstrap(qp)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{parbootstrap} creates bootstrap samples of the parameters.
}

