% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parci.R
\name{parci}
\alias{parci}
\title{Calculates the confidence intervals for the parameters
\code{parci} calculates the confidence intervals for the parameters}
\usage{
parci(par, parbootstrap, ci)
}
\description{
Calculates the confidence intervals for the parameters
\code{parci} calculates the confidence intervals for the parameters
}
\keyword{internal}
