% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parini.R
\name{parini}
\alias{parini}
\title{Calculates the initial parameters}
\usage{
parini(averages, funname_df, x, guess, lapses, pariniset, parini)
}
\description{
Calculates the initial parameters
}
\keyword{internal}
