% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pariniold.R
\name{pariniold}
\alias{pariniold}
\title{Calculates some initial parameters
\code{pariniols} calculates some initial parameters}
\usage{
pariniold(d, x, k, n, guess, lapses, psyfun)
}
\description{
Calculates some initial parameters
\code{pariniols} calculates some initial parameters
}
\keyword{internal}
