% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/one_bootstrap.R
\docType{data}
\name{parn}
\alias{parn}
\title{Performs the bootstrap for one condition
\code{one_bootstrap} performs the bootstrap for one condition}
\format{\preformatted{ chr "No te quejes"
}}
\usage{
parn
}
\description{
Performs the bootstrap for one condition
\code{one_bootstrap} performs the bootstrap for one condition
}
\keyword{internal}

