% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotcurves.R
\name{plotcurves}
\alias{plotcurves}
\title{plotcurves}
\usage{
plotcurves(qp, panel = NULL, xpanel = NULL, ypanel = NULL, color = NULL,
  averages = T, curves = T, thresholds = T, ci = T)
}
\description{
plotcurves
}

