% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotpar_.R
\name{plotpar_}
\alias{plotpar_}
\title{Plot the values of the parameters}
\usage{
plotpar_(qp, x = NULL, panel = NULL, xpanel = NULL, ypanel = NULL,
  color = NULL, geom = "bar", ci = T)
}
\description{
\code{plotpar_} is the standard evaluation SE function associated
to the non-standard evaluation NSE function \code{plotpar}.
\href{http://adv-r.had.co.nz/Computing-on-the-language.html}{SE functions can be more easily called from other functions.}
In SE functions, you need to quote the names of the variables.
}
\examples{
#' library(MPDiR) # contains the Vernier data
data(Vernier) # ?Venier for the reference
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq))

plotpar_(fit, x = 'WaveForm')
plotpar_(fit, xpanel = 'Direction')
plotpar_(fit, color = 'Direction')
plotpar_(fit, color = 'Direction', ypanel = 'WaveForm', geom = 'point')
}
\seealso{
\code{\link{plotpar}}
}

