% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotthresholds_.R
\name{plotthresholds_}
\alias{plotthresholds_}
\title{plotthresholds_}
\usage{
plotthresholds_(qp, x = NULL, panel = NULL, xpanel = NULL,
  ypanel = NULL, color = NULL, geom = "bar", ci = T)
}
\description{
plotthresholds_
}

