% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psyfunguesslapses_df.R
\name{psyfunguesslapses_df}
\alias{psyfunguesslapses_df}
\title{Creates the psychometric function
\code{psyfunguesslapses_df} creates the psychometric function}
\usage{
psyfunguesslapses_df(functions_df, guess, lapses)
}
\description{
Creates the psychometric function
\code{psyfunguesslapses_df} creates the psychometric function
}
\keyword{internal}
