% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickpsy.R
\docType{data}
\name{qpdat}
\alias{qpdat}
\title{Data set for demonstration}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6240 rows and 8 columns.}
\usage{
qpdat
}
\description{
It is part of the data associated with the paper 'Motion signal and
the perceived positions of moving objects'.
}
\references{
Linares, D., López-Moliner, J., & Johnston, A. (2007). Motion
signal and the perceived positions of moving objects. Journal of Vision,
7(7), 1.
}
\keyword{datasets}

