% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickpsy_.R
\name{quickpsy_}
\alias{quickpsy_}
\title{Deprecated. If you need non-standard evaluation look at
\href{http://dplyr.tidyverse.org/articles/programming.html#warm-up}{this.}}
\usage{
quickpsy_()
}
\description{
Deprecated. If you need non-standard evaluation look at
\href{http://dplyr.tidyverse.org/articles/programming.html#warm-up}{this.}
}
\seealso{
\code{\link{quickpsy}}
}
