% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickpsys_.R
\name{quickpsys_}
\alias{quickpsys_}
\title{Fits psychometric functions that share parameters}
\usage{
quickpsys_(d, x = "x", k = "k", n = "n", grouping, random, within,
  between, xmin = NULL, xmax = NULL, log = FALSE, fun = NULL,
  parini = NULL, prob = NULL, thresholds = T, bootstrap = "parametric",
  B = 100, ci = 0.95, optimization = "optim")
}
\arguments{
\item{d}{Data frame with the results of a Yes-No experiment to fit.
It should have a
\href{http://vita.had.co.nz/papers/tidy-data.html}{tidy} form in which
each column corresponds to a variable and each row is an observation.}

\item{x}{Name of the explanatory variable.}

\item{k}{Name of the response variable. The response variable could be the
number of trials in which a yes-type response was given or a vector of 0s
(or -1s; no-type response) and 1s (yes-type response) indicating the
response on each trial.}

\item{n}{Only necessary if \code{k} refers to the number of trials
in which a yes-type response was given. It corresponds to the name of the
variable indicating the total number of trials.}

\item{grouping}{Name of the grouping variables. It should be specified as
\code{grouping = .(variable_name1, variable_name2)}.}

\item{random}{Name of the random variable. It should be specified as
\code{random = .(variable_name1, variable_name2)}. In the current version
of quickpsy, the random variable has not special treatment. It does the
same as \code{grouping}.}

\item{within}{Name of the within variable. It should be specified as
\code{within = .(variable_name1, variable_name2)}. In the current version
of quickpsy, the within variable has not special treatment. It does the
same as \code{grouping}.}

\item{between}{Name of the between variable.  It should be specified as
\code{between = .(variable_name1, variable_name2)}. In the current version
of quickpsy, the between variable has not special treatment. It does the
same as \code{grouping}.}

\item{xmin}{Minimum value of the explanatory variable for which the curves
should be calculated (the default is the minimum value of the explanatory
variable).}

\item{xmax}{Maximum value of the explanatory variable for which the curves
should be calculated (the default is the maximum value of the explanatory
variable).}

\item{log}{If \code{TRUE}, the logarithm of the explanatory variable is used
to fit the curves (default is \code{FALSE}).}

\item{fun}{Vector of string names of the functions introduced by the user}

\item{parini}{Initial parameters. It could be a vector or a list of the form
\code{list(c(par1min, par2max), c(par2min, par2max))} to
constraint the lower and upper bounds of the parameters (when
\code{optimization = 'DE'}, parini should be also a list).}

\item{prob}{Probability to calculate the threshold (default is
\code{guess + .5 * (1 - guess)}).}

\item{thresholds}{If \code{FALSE}, thresholds are not calculated
(default is \code{TRUE}).}

\item{bootstrap}{\code{'parametric'} performs parametric bootstrap;
\code{'nonparametric'} performs non-parametric bootstrap;
\code{'none'} does not perform bootstrap (default is \code{'parametric'}).}

\item{B}{number of bootstrap samples (default is 100 ONLY).}

\item{ci}{Confidence intervals level based on percentiles (default is .95).}

\item{optimization}{Method used for optimizization. The default is 'optim' which uses
the \code{optim} function. It can also be \code{'DE'} which uses de function
\code{DEoptim} from the package DEoptim, which performs differential
evolution optimization. By using \code{DEoptim}, it is less likely that the
optimization finishes in a local minimum, but the optimization is slow.
When \code{'DE'} is used, \code{parini} should be specified as a list with
lower and upper bounds.}
}
\description{
\code{quickpsys_} is the standard evaluation SE function associated
to the non-standard evaluation NSE function \code{quickpsys}.
\href{http://adv-r.had.co.nz/Computing-on-the-language.html}{SE functions can be more easily called from other functions.}
In SE functions, you need to quote the names of the variables.
}
\examples{
library(MPDiR) # contains the Vernier data
fit <- quickpsy_(Vernier, 'Phaseshift', 'NumUpward', 'N',
                grouping = c('Direction', 'WaveForm', 'TempFreq'), B = 20)
plotcurves(fit)
}
\seealso{
\code{\link{quickpsy}}
}

