% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.quickpsy}
\alias{summary.quickpsy}
\title{Plot the parameters and its confidence intervals
\code{summary} Plot the parameters and its confidence intervals}
\usage{
\method{summary}{quickpsy}(object, ...)
}
\arguments{
\item{object}{An object for which a summary is desired.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Plot the parameters and its confidence intervals
\code{summary} Plot the parameters and its confidence intervals
}

