% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Calculates the thresholds}
\usage{
threshold(par, curves, prob, log, funname, guess, lapses)
}
\arguments{
\item{prob}{Probability to calculate the threshold.}

\item{log}{Use \code{TRUE}, if the logarithm of the independent variable
has been used to fit the curves (default is \code{FALSE}).}

\item{qp}{output from quickpsy}
}
\description{
Calculates the thresholds
}
\keyword{Calculates}
\keyword{\code{thresholds}}
\keyword{internal}
\keyword{the}
\keyword{thresholds}
