% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdcomparisons.R
\name{thresholdcomparisons}
\alias{thresholdcomparisons}
\title{Pair comparisons of the thresholds using bootstrap
\code{thresholdscomparisons} Calculates the bootstrap confidence intervals for the
difference in the thresholds for two groups for all possible pairs
of groups}
\usage{
thresholdcomparisons(thresholds_dif, thresholds_difbootstrap, ci)
}
\description{
Pair comparisons of the thresholds using bootstrap
\code{thresholdscomparisons} Calculates the bootstrap confidence intervals for the
difference in the thresholds for two groups for all possible pairs
of groups
}
\keyword{internal}
