% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdsbootstrap.R
\name{thresholdsbootstrap}
\alias{thresholdsbootstrap}
\title{Calculates bootstrap thresholds
\code{thresholdsbootstrap} calculates bootstrap thresholds}
\usage{
thresholdsbootstrap(qp, prob = NULL, log = F)
}
\description{
Calculates bootstrap thresholds
\code{thresholdsbootstrap} calculates bootstrap thresholds
}
\keyword{internal}

