% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshodsci.R
\name{thresholdsci}
\alias{thresholdsci}
\title{Calculates the confidence intervals for the thresholds
\code{thresholdsci} calculates the confidence intervals for the thresholds}
\usage{
thresholdsci(thresholds, thresholdsbootstrap, ci)
}
\description{
Calculates the confidence intervals for the thresholds
\code{thresholdsci} calculates the confidence intervals for the thresholds
}
\keyword{internal}
