% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_seq.R
\name{x_seq}
\alias{x_seq}
\title{Creates sequences of x's}
\usage{
x_seq(limits, x, grouping)
}
\arguments{
\item{limits}{The \code{limits} data frame from quickpsy.}

\item{x}{The \code{x} data frame from quickpsy.}

\item{grouping}{The \code{grouping} data frame from quickpsy.}
}
\description{
Creates sequences of x's
}
