% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypred.R
\name{ypred}
\alias{ypred}
\title{Predicted probabilities}
\usage{
ypred(
  averages,
  param,
  psych_fun,
  x_str,
  log,
  grouping,
  grouping_without_fun,
  grouping_fun
)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{ypred} calculates the predicted probabilities at the values of the
explanatory variable.
}
\examples{
library(MPDiR) # contains the Vernier data
data(Vernier) # ?Venier for the reference
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq), B = 20)
ypred(fit)
}
