% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_nll.R
\name{create_nll}
\alias{create_nll}
\title{It creates from the data the negative log-likelihood as a function of the parameters.}
\usage{
create_nll(d, x, k, n, psy_fun)
}
\arguments{
\item{d}{The data frame containing the data.}

\item{x}{Name of the independent variable.}

\item{k}{Number of 'yes' or correct responses.}

\item{n}{Number of trials.}

\item{psy_fun}{Function to create the negative log-likelihood.}
}
\value{
It returns a function. The negative log-likelihood as a function of the parameters (p).
}
\description{
It creates from the data the negative log-likelihood as a function of the parameters.
}

