% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_nll.R
\name{create_nll}
\alias{create_nll}
\title{Creates the negative log-likelihood function
\code{create_nll} Creates the negative log-likelihood function}
\usage{
create_nll(d, x, k, n, psyfunguesslapses)
}
\description{
Creates the negative log-likelihood function
\code{create_nll} Creates the negative log-likelihood function
}
\keyword{internal}

