% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_psy_fun.R
\name{create_psy_fun}
\alias{create_psy_fun}
\title{Creates the psychometric function
\code{create_psy_fun} creates the psychometric function}
\usage{
create_psy_fun(psy_fun, guess, lapses)
}
\description{
Creates the psychometric function
\code{create_psy_fun} creates the psychometric function
}
\keyword{internal}

