% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fit_main_glm}
\alias{fit_main_glm}
\title{fit_main_glm}
\usage{
fit_main_glm(d, x, k, n, psy_fun, psy_fun_name, pini, guess, lapses)
}
\arguments{
\item{d}{The data frame containing the data.}

\item{x}{Name of the independent variable.}

\item{k}{Number of 'yes' or correct responses.}

\item{n}{Number of trials.}

\item{psy_fun}{Function to create the negative log-likelihood.}

\item{pini}{Initial parameters.}
}
\description{
fit_main_glm
}

