% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitpsy.R
\name{fitpsy}
\alias{fitpsy}
\title{Fits the curve
\code{fitpsy} fits de curve}
\usage{
fitpsy(d, x, k, n, random, within, between, grouping, xmin, xmax, log, funname,
  parini, pariniset, guess, lapses, optimization)
}
\description{
Fits the curve
\code{fitpsy} fits de curve
}
\keyword{internal}

