% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/limits.R
\name{limits}
\alias{limits}
\title{Creates the limits
\code{limits} creates the limits}
\usage{
limits(d, x, xmin, xmax, log)
}
\description{
Creates the limits
\code{limits} creates the limits
}
\keyword{internal}

