% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliks.R
\name{logliks}
\alias{logliks}
\title{Calculates the loglikelihoods
\code{logliks} calculates the loglikelihoods.}
\usage{
logliks(nll_fun, param, grouping_without_fun)
}
\description{
Calculates the loglikelihoods
\code{logliks} calculates the loglikelihoods.
}
