% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikssaturated.R
\name{loglikssaturated}
\alias{loglikssaturated}
\title{Calculates the saturated loglikelihoods}
\usage{
loglikssaturated(qp)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{loglikssaturated} calculates the saturated loglikelihoods.
}
\examples{
library(MPDiR) # contains the Vernier data
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq), B = 20)
loglikssaturated(fit)
}

