% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikssaturated.R
\name{loglikssaturated}
\alias{loglikssaturated}
\title{Calculates the saturated loglikelihoods
\code{logliks} calculates the saturatedloglikelihoods.}
\usage{
loglikssaturated(nll_fun_saturated, averages, grouping_without_fun)
}
\description{
Calculates the saturated loglikelihoods
\code{logliks} calculates the saturatedloglikelihoods.
}
