% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_fun.R
\name{nll_fun}
\alias{nll_fun}
\title{\code{nll_fun} Creates the negative log-likelihood function}
\usage{
nll_fun(averages, psych_fun, x_str, grouping_without_fun, grouping_fun)
}
\description{
\code{nll_fun} Creates the negative log-likelihood function
}
\keyword{internal}
