% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_fun.R
\name{nll_fun}
\alias{nll_fun}
\title{\code{nll_fun} Creates the negative log-likelihood function}
\usage{
nll_fun(averages, psych_fun, x_str, grouping_without_fun, grouping_fun)
}
\arguments{
\item{averages}{The \code{averages} data frame from quickpsy.}

\item{psych_fun}{The \code{psych_fun} data frame from quickpsy.}

\item{x_str}{The \code{x_str} data frame from quickpsy.}

\item{grouping_without_fun}{The \code{grouping_without_fun} data frame from quickpsy.}

\item{grouping_fun}{The \code{grouping_fun} data frame from quickpsy.}
}
\description{
\code{nll_fun} Creates the negative log-likelihood function
}
\keyword{internal}
