% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/one_parameters.R
\name{one_parameters}
\alias{one_parameters}
\title{one_parameters}
\usage{
one_parameters(d, x, k, n, psyfunguesslapses, funname, pini, guess, lapses,
  pini2, groups)
}
\description{
one_parameters
}

