% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Calculates the parameters
\code{parameters} calculates the parameters}
\usage{
parameters(averages, parini, psyfunguesslapses_df, funname_df, x, guess, lapses)
}
\description{
Calculates the parameters
\code{parameters} calculates the parameters
}
\keyword{internal}
