% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Calculates the parameters
\code{parameters} calculates the parameters}
\usage{
parameters(d, x, k, n, psyfunguesslapses, funname, parini, pariniset, guess,
  lapses, optimization, groups)
}
\description{
Calculates the parameters
\code{parameters} calculates the parameters
}
\keyword{internal}

