% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parcomparisons.R
\name{parcomparisons}
\alias{parcomparisons}
\title{Pair comparisons of the parameters using bootstrap
\code{parcomparisons} Calculates the bootstrap confidence intervals for the
difference in the parameters for two groups for all possible pairs
of groups}
\usage{
parcomparisons(qp, ci = 0.95)
}
\description{
Pair comparisons of the parameters using bootstrap
\code{parcomparisons} Calculates the bootstrap confidence intervals for the
difference in the parameters for two groups for all possible pairs
of groups
}
\keyword{internal}

