% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parini.R
\name{parini}
\alias{parini}
\title{Calculates some initial parameters
\code{parini} calculates some initial parameters}
\usage{
parini(d, x, k, n, guess, lapses, psyfun)
}
\description{
Calculates some initial parameters
\code{parini} calculates some initial parameters
}
\keyword{internal}

