% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_bootstrapav.R
\docType{data}
\name{parn}
\alias{parn}
\title{Creates one bootstrap sample
\code{one_bootstrapav} creates one bootstrap sample}
\format{An object of class \code{character} of length 1.}
\usage{
parn
}
\description{
Creates one bootstrap sample
\code{one_bootstrapav} creates one bootstrap sample
}
\keyword{internal}

