% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_bootstrap.R
\docType{data}
\name{parn}
\alias{parn}
\title{Performs the bootstrap for one condition
\code{one_bootstrap} performs the bootstrap for one condition}
\format{An object of class \code{character} of length 1.}
\usage{
parn
}
\description{
Performs the bootstrap for one condition
\code{one_bootstrap} performs the bootstrap for one condition
}
\keyword{internal}

