% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotcurves_.R
\name{plotcurves_}
\alias{plotcurves_}
\title{plotcurves_}
\usage{
plotcurves_(qp, panel = NULL, xpanel = NULL, ypanel = NULL,
  color = NULL, averages = T, curves = T, thresholds = T)
}
\description{
plotcurves_
}

