% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotthresholds.R
\name{plotthresholds}
\alias{plotthresholds}
\title{plotthresholds}
\usage{
plotthresholds(qp, x = NULL, panel = NULL, xpanel = NULL, ypanel = NULL,
  color = NULL, geom = "bar")
}
\description{
plotthresholds
}

