% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qpdat}
\alias{qpdat}
\title{Data set for demonstration}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1280 rows and 5 columns.
}
\usage{
qpdat
}
\description{
It is part of the data associated with the paper "Linares, D.,
& Lopez-Moliner, J. (2006). Perceptual asynchrony between color and motion
with a single direction change. Journal of Vision, 6(9), 10-10".
}
\keyword{datasets}
