% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quickpsy.R
\docType{data}
\name{qpdat}
\alias{qpdat}
\title{Data set for demonstration}
\description{
It is part of the data associated with the paper 'Motion signal and
the perceived positions of moving objects'.
}
\references{
Linares, D., López-Moliner, J., & Johnston, A. (2007). Motion
signal and the perceived positions of moving objects. Journal of Vision,
7(7), 1.
}

