% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quickpsy.R
\name{quickpsy}
\alias{quickpsy}
\title{quickpsy}
\usage{
quickpsy(d, x, k, n = NULL, random, within, between, xmin = NULL,
  xmax = NULL, log = F, psyfun = cum_normal_fun, pini = NULL,
  guess = 0, lapses = 0, DE = F, pini2 = NULL, threprob = NULL,
  curv = T, thre = T, plotcurves = T)
}
\description{
quickpsy
}

