% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickreadfiles.R
\name{quickreadfiles}
\alias{quickreadfiles}
\title{Reads several files}
\usage{
quickreadfiles(path = getwd(), extension = "txt", ...)
}
\arguments{
\item{path}{Path of the file (default is the working directory).}

\item{extension}{Specify whether the file extension is 'txt' or 'csv'.}

\item{...}{arguments of the form name_var = c('value1', 'value2',..).
A new column with variable name name_var is addes to the data frame.}
}
\description{
\code{quickreadfiles} builts a data frame from several txt files. It
assumes that in each file, the first row has the names of the variables.
}
\examples{
# download the 3 files in
# https://github.com/danilinares/quickpsy/tree/master/inst/extdata/example1
# and add them to your working directory
# dat <- quickreadfiles(subject = c('aa', 'bb', 'cc'), session = c('1', '2'))
# fit <- quickpsy(dat, phase, resp, grouping=.(subject), lapses = T, guess = T)
# plotcurves(fit)
}
