% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdsci.R
\name{thresholdsci}
\alias{thresholdsci}
\title{Calculates the confidence intervals for the thresholds
\code{thresholdsci} calculates the confidence intervals for the thresholds}
\usage{
thresholdsci(qp, ci = 0.95, method = "percent")
}
\description{
Calculates the confidence intervals for the thresholds
\code{thresholdsci} calculates the confidence intervals for the thresholds
}
\keyword{internal}

