% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypred.R
\name{ypred}
\alias{ypred}
\title{Predicted probabilities}
\usage{
ypred(
  averages,
  param,
  psych_fun,
  x_str,
  log,
  grouping,
  grouping_without_fun,
  grouping_fun,
  funname,
  guess,
  lapses
)
}
\arguments{
\item{averages}{The \code{"averages"} data frame from quickpsy.}

\item{param}{The \code{"param"} data frame from quickpsy.}

\item{psych_fun}{The \code{"psych_fun"} data frame from quickpsy.}

\item{x_str}{The \code{"x_str"} data frame from quickpsy.}

\item{log}{The \code{"log"} data frame from quickpsy.}

\item{grouping}{The \code{"grouping"} data frame from quickpsy.}

\item{grouping_without_fun}{The \code{"grouping_without_fun"} data frame from quickpsy.}

\item{grouping_fun}{The \code{"grouping_fun"} data frame from quickpsy.}

\item{funname}{The \code{funname} data frame from quickpsy.}

\item{guess}{Argument \code{guess}}

\item{lapses}{Argument \code{lapses}}
}
\description{
\code{ypred} calculates the predicted probabilities at the values of the
explanatory variable.
}
