% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ypred.R
\name{ypred}
\alias{ypred}
\title{Predicted probabilities}
\usage{
ypred(qp)
}
\arguments{
\item{qp}{output from quickpsy}
}
\description{
\code{ypred} calculates the predicted probabilities at the values of the
explanatory variable.
}
\examples{
library(MPDiR) # contains the Vernier data
data(Vernier) # ?Venier for the reference
fit <- quickpsy(Vernier, Phaseshift, NumUpward, N,
                grouping = .(Direction, WaveForm, TempFreq))
ypred(fit)
}

